fix(Java): Changed FilterableNotificationExtractor from an interface to a class and added generics to NotificationExtractor.
