+++ processor-api/src/main/java/com/synopsys/integration/alert/processor/api/filter/FilterableNotificationExtractor.java
@@ -22,12 +22,55 @@
  */
 package com.synopsys.integration.alert.processor.api.filter;
 
+import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
+
 import com.synopsys.integration.alert.common.rest.model.AlertNotificationModel;
+import com.synopsys.integration.alert.common.util.DataStructureUtils;
+import com.synopsys.integration.alert.processor.api.filter.extractor.NotificationExtractor;
 import com.synopsys.integration.alert.processor.api.filter.model.FilterableNotificationWrapper;
+import com.synopsys.integration.blackduck.api.manual.enumeration.NotificationType;
 
-public interface FilterableNotificationExtractor {
-    Optional<FilterableNotificationWrapper<?>> wrapNotification(AlertNotificationModel notification);
+@Component
+public class FilterableNotificationExtractor {
+    private final Logger logger = LoggerFactory.getLogger(FilterableNotificationExtractor.class);
 
+    private Map<NotificationType, NotificationExtractor> notificationExtractors;
+
+    @Autowired
+    public FilterableNotificationExtractor(List<NotificationExtractor> notificationExtractors) {
+        this.notificationExtractors = DataStructureUtils.mapToValues(notificationExtractors, NotificationExtractor::getNotificationType);
+    }
+
+    public final Optional<FilterableNotificationWrapper<?>> wrapNotification(AlertNotificationModel notification) {
+        String notificationTypeString = notification.getNotificationType();
+        NotificationType notificationType;
+        try {
+            notificationType = Enum.valueOf(NotificationType.class, notificationTypeString);
+        } catch (IllegalArgumentException e) {
+            logger.warn("Notification did not match any existing notification type: {}", notificationTypeString);
+            return Optional.empty();
+        }
+
+        Optional<NotificationExtractor> notificationExtractorOptional = getNotificationExtractor(notificationType);
+        if (notificationExtractorOptional.isPresent()) {
+            NotificationExtractor notificationExtractor = notificationExtractorOptional.get();
+
+            return Optional.of(notificationExtractor.convertToFilterableNotificationWrapper(notification));
+        }
+
+        logger.warn("Did not find extractor for notification type: {}", notificationTypeString);
+        return Optional.empty();
+    }
+
+    private Optional<NotificationExtractor> getNotificationExtractor(NotificationType notificationType) {
+        NotificationExtractor notificationExtractor = notificationExtractors.get(notificationType);
+        return Optional.ofNullable(notificationExtractor);
+    }
 }
