+++ processor-api/src/main/java/com/synopsys/integration/alert/processor/api/filter/extractor/NotificationExtractor.java
@@ -24,9 +24,10 @@ package com.synopsys.integration.alert.processor.api.filter.extractor;
 
 import com.synopsys.integration.alert.common.rest.model.AlertNotificationModel;
 import com.synopsys.integration.alert.processor.api.filter.model.FilterableNotificationWrapper;
+import com.synopsys.integration.blackduck.api.manual.component.NotificationContentComponent;
 import com.synopsys.integration.blackduck.api.manual.enumeration.NotificationType;
 
-public abstract class NotificationExtractor {
+public abstract class NotificationExtractor<T extends NotificationContentComponent> {
     private NotificationType notificationType;
 
     public NotificationExtractor(NotificationType notificationType) {
@@ -37,6 +38,6 @@ public abstract class NotificationExtractor {
         return notificationType;
     }
 
-    public abstract FilterableNotificationWrapper convertToFilterableNotificationWrapper(AlertNotificationModel alertNotificationModel);
+    public abstract FilterableNotificationWrapper<T> convertToFilterableNotificationWrapper(AlertNotificationModel alertNotificationModel);
 
 }
