+++ processor-api/src/main/java/com/synopsys/integration/alert/processor/api/filter/extractor/VulnerabilityNotificationExtractor.java
@@ -41,7 +41,7 @@ import com.synopsys.integration.blackduck.api.manual.component.VulnerabilitySour
 import com.synopsys.integration.blackduck.api.manual.enumeration.NotificationType;
 
 @Component
-public class VulnerabilityNotificationExtractor extends NotificationExtractor {
+public class VulnerabilityNotificationExtractor extends NotificationExtractor<VulnerabilityNotificationContent> {
     private Gson gson;
 
     @Autowired
@@ -51,7 +51,7 @@ public class VulnerabilityNotificationExtractor extends NotificationExtractor {
     }
 
     @Override
-    public FilterableNotificationWrapper convertToFilterableNotificationWrapper(AlertNotificationModel alertNotificationModel) {
+    public FilterableNotificationWrapper<VulnerabilityNotificationContent> convertToFilterableNotificationWrapper(AlertNotificationModel alertNotificationModel) {
         VulnerabilityNotificationContent vulnerabilityNotificationContent = gson.fromJson(alertNotificationModel.getContent(), VulnerabilityNotificationContent.class);
 
         List<String> affectedProjectNames = vulnerabilityNotificationContent.getAffectedProjectVersions()
