+++ processor-api/src/test/java/com/synopsys/integration/alert/processor/api/filter/extractor/VulnerabilityNotificationExtractorTest.java
@@ -32,7 +32,7 @@ public class VulnerabilityNotificationExtractorTest {
         VulnerabilityNotificationExtractor vulnerabilityNotificationExtractor = new VulnerabilityNotificationExtractor(gson);
 
         AlertNotificationModel alertNotificationModel = createAlertNotificationModel(gson.toJson(content));
-        FilterableNotificationWrapper filterableNotificationWrapper = vulnerabilityNotificationExtractor.convertToFilterableNotificationWrapper(alertNotificationModel);
+        FilterableNotificationWrapper<VulnerabilityNotificationContent> filterableNotificationWrapper = vulnerabilityNotificationExtractor.convertToFilterableNotificationWrapper(alertNotificationModel);
 
         String notificationType = filterableNotificationWrapper.extractNotificationType();
